test(summary): scope projection budget fixtures
